# openapi.json
{
  "info": {
    "title": "Atiyeh Holdings",
    "version": "0.0.1"
  },
  "openapi": "3.0.3",
  "paths": {
    "/users/{userId}": {
      "get": {
        "operationId": "UserApiSpec_get_/{userId}",
        "tags": [],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/default"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "UserApiSpec_put_/{userId}",
        "tags": [],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "default": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "_id": {
            "description": "Add an optional _id field to an object shaped type",
            "$ref": "#/components/schemas/OptionalId_ObjectId_"
          },
          "houses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/default_1"
            }
          }
        },
        "additionalProperties": false,
        "required": [
          "email",
          "name"
        ]
      },
      "OptionalId_ObjectId_": {
        "description": "Add an optional _id field to an object shaped type",
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "_bsontype": {
            "type": "string",
            "enum": [
              "ObjectId"
            ]
          },
          "equals": {
            "description": "Compares the equality of this ObjectId with `otherID`.",
            "type": "object",
            "additionalProperties": false
          },
          "toString": {
            "description": "Converts the id into a 24 character hex string for printing, unless encoding is provided.",
            "type": "object",
            "additionalProperties": false
          },
          "inspect": {
            "description": "Converts to a string representation of this Id.",
            "type": "object",
            "additionalProperties": false
          },
          "id": {
            "description": "The ObjectId bytes",
            "type": "object",
            "additionalProperties": false,
            "patternProperties": {
              "^[0-9]+$": {
                "type": "number"
              }
            },
            "properties": {
              "BYTES_PER_ELEMENT": {
                "type": "number"
              },
              "buffer": {
                "$ref": "#/components/schemas/ArrayBufferLike"
              },
              "byteLength": {
                "type": "number"
              },
              "byteOffset": {
                "type": "number"
              },
              "length": {
                "type": "number"
              },
              "__@toStringTag@4727": {
                "type": "string",
                "enum": [
                  "Uint8Array"
                ]
              }
            },
            "required": [
              "BYTES_PER_ELEMENT",
              "__@toStringTag@4727",
              "buffer",
              "byteLength",
              "byteOffset",
              "length"
            ]
          },
          "toHexString": {
            "description": "Returns the ObjectId id as a 24 lowercase character hex string representation",
            "type": "object",
            "additionalProperties": false
          },
          "toJSON": {
            "description": "Converts to its JSON the 24 character hex string representation.",
            "type": "object",
            "additionalProperties": false
          },
          "getTimestamp": {
            "description": "Returns the generation date (accurate up to the second) that this ID was generated.",
            "type": "object",
            "additionalProperties": false
          },
          "_id": {
            "description": "Given an object shaped type, return the type of the _id field or default to ObjectId",
            "$ref": "#/components/schemas/ObjectId"
          }
        },
        "required": [
          "_bsontype",
          "equals",
          "getTimestamp",
          "id",
          "inspect",
          "toHexString",
          "toJSON",
          "toString"
        ]
      },
      "ArrayBufferLike": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/ArrayBuffer"
          },
          {
            "$ref": "#/components/schemas/SharedArrayBuffer"
          }
        ]
      },
      "ArrayBuffer": {
        "type": "object",
        "properties": {
          "byteLength": {
            "type": "number"
          },
          "__@toStringTag@4727": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "__@toStringTag@4727",
          "byteLength"
        ]
      },
      "SharedArrayBuffer": {
        "type": "object",
        "properties": {
          "byteLength": {
            "type": "number"
          },
          "__@species@6399": {
            "$ref": "#/components/schemas/SharedArrayBuffer"
          },
          "__@toStringTag@4727": {
            "type": "string",
            "enum": [
              "SharedArrayBuffer"
            ]
          }
        },
        "additionalProperties": false,
        "required": [
          "__@species@6399",
          "__@toStringTag@4727",
          "byteLength"
        ]
      },
      "ObjectId": {
        "description": "A class representation of the BSON ObjectId type.",
        "type": "object",
        "properties": {
          "_bsontype": {
            "type": "string",
            "enum": [
              "ObjectId"
            ]
          },
          "id": {
            "description": "The ObjectId bytes",
            "type": "object",
            "additionalProperties": false,
            "patternProperties": {
              "^[0-9]+$": {
                "type": "number"
              }
            },
            "properties": {
              "BYTES_PER_ELEMENT": {
                "type": "number"
              },
              "buffer": {
                "$ref": "#/components/schemas/ArrayBufferLike"
              },
              "byteLength": {
                "type": "number"
              },
              "byteOffset": {
                "type": "number"
              },
              "length": {
                "type": "number"
              },
              "__@toStringTag@4727": {
                "type": "string",
                "enum": [
                  "Uint8Array"
                ]
              }
            },
            "required": [
              "BYTES_PER_ELEMENT",
              "__@toStringTag@4727",
              "buffer",
              "byteLength",
              "byteOffset",
              "length"
            ]
          }
        },
        "additionalProperties": false,
        "required": [
          "_bsontype",
          "id"
        ]
      },
      "default_1": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/default"
            }
          },
          "details": {
            "type": "object",
            "properties": {
              "bathrooms": {
                "type": "number"
              },
              "bedrooms": {
                "type": "number"
              },
              "currentValue": {
                "type": "number"
              },
              "purchasePrice": {
                "type": "number"
              },
              "sqft": {
                "type": "number"
              }
            },
            "additionalProperties": false,
            "required": [
              "bathrooms",
              "bedrooms",
              "currentValue",
              "purchasePrice",
              "sqft"
            ]
          },
          "id": {
            "description": "Add an optional _id field to an object shaped type",
            "$ref": "#/components/schemas/OptionalId_ObjectId_"
          },
          "lease": {
            "type": "object",
            "properties": {
              "deposit": {
                "type": "number"
              },
              "end": {
                "type": "string",
                "format": "date-time"
              },
              "rentPrice": {
                "type": "number"
              },
              "start": {
                "type": "string",
                "format": "date-time"
              },
              "tenant": {}
            },
            "additionalProperties": false,
            "required": [
              "deposit",
              "end",
              "rentPrice",
              "start",
              "tenant"
            ]
          }
        },
        "additionalProperties": false,
        "required": [
          "address",
          "name",
          "users"
        ]
      }
    }
  }
}